requirements.txt
streamlit_lottie
bs4